+++ pyanalyze/annotations.py
@@ -24,7 +24,6 @@ show errors.
 
 """
 from dataclasses import dataclass, InitVar, field
-import mypy_extensions
 import typing_extensions
 import typing
 import typing_inspect
@@ -35,7 +34,7 @@ from collections.abc import Callable, Iterable
 from typing import (
     Any,
     Container,
-    Dict,
+    NamedTuple,
     cast,
     TypeVar,
     ContextManager,
@@ -59,6 +58,7 @@ from .extensions import (
 )
 from .find_unused import used
 from .signature import SigParameter, Signature
+from .safe import is_typing_name, is_instance_of_typing_name
 from .value import (
     AnnotatedValue,
     AnySource,
@@ -438,6 +438,8 @@ def _type_from_runtime(val: Any, ctx: Context, is_typeddict: bool = False) -> Va
         origin = get_origin(val)
         if isinstance(origin, type):
             return _maybe_typed_value(origin, ctx)
+        elif val is NamedTuple:
+            return TypedValue(tuple)
         ctx.show_error(f"Invalid type annotation {val}")
         return AnyValue(AnySource.error)
 
@@ -494,6 +496,11 @@ def _type_from_value(value: Value, ctx: Context, is_typeddict: bool = False) -> 
                     root_type.typ,
                     [_type_from_value(member, ctx) for member in value.members],
                 )
+        if isinstance(value.root, TypedValue) and isinstance(value.root.typ, str):
+            return GenericValue(
+                value.root.typ, [_type_from_value(elt, ctx) for elt in value.members]
+            )
+
         if not isinstance(value.root, KnownValue):
             ctx.show_error(f"Cannot resolve subscripted annotation: {value.root}")
             return AnyValue(AnySource.error)
@@ -595,6 +602,9 @@ def _type_from_value(value: Value, ctx: Context, is_typeddict: bool = False) -> 
             ctx.show_error(f"Unrecognized subscripted annotation: {root}")
             return AnyValue(AnySource.error)
     elif isinstance(value, AnyValue):
+        return value
+    elif isinstance(value, TypedValue) and isinstance(value.typ, str):
+        # Synthetic type
         return value
     else:
         ctx.show_error(f"Unrecognized annotation {value}")
@@ -759,37 +769,6 @@ class _Visitor(ast.NodeVisitor):
             return TypedValue(func.val)
         else:
             return None
-
-
-def is_typing_name(obj: object, name: str) -> bool:
-    objs = _fill_typing_name_cache(name)
-    for typing_obj in objs:
-        if obj is typing_obj:
-            return True
-    return False
-
-
-def is_instance_of_typing_name(obj: object, name: str) -> bool:
-    objs = _fill_typing_name_cache(name)
-    return isinstance(obj, objs)
-
-
-_typing_name_cache: Dict[str, Tuple[Any, ...]] = {}
-
-
-def _fill_typing_name_cache(name: str) -> Tuple[Any, ...]:
-    try:
-        return _typing_name_cache[name]
-    except KeyError:
-        objs = []
-        for mod in (typing, typing_extensions, mypy_extensions):
-            try:
-                objs.append(getattr(mod, name))
-            except AttributeError:
-                pass
-        objs_tuple = tuple(objs)
-        _typing_name_cache[name] = objs_tuple
-        return objs_tuple
 
 
 def _value_of_origin_args(